Refactor/task 11 refactorizar

+++ Backend/src/main/java/com/HolosINC/Holos/Category/Category.java
@@ -6,9 +6,8 @@ import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.SequenceGenerator;
 import jakarta.persistence.Table;
-import lombok.Data;
-//import com.HolosINC.Holos.Work;
 
+import lombok.Data;
 
 @Data
 @Entity
@@ -18,11 +17,8 @@ public class Category {
     @Id
     @SequenceGenerator(name = "category_seq", sequenceName = "category_sequence", initialValue = 100)
     @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "category_seq")
-    private Integer id;
+    private Long id;
 
     private String name;
     private String description;
-    
-    //@OneToMany(mappedBy = "category")
-    //private Set<Work> works;
 }
